% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_fun.R
\name{get_TPMlike_counts}
\alias{get_TPMlike_counts}
\title{Normalizing raw ATAC-Seq counts using a TPM-like apporach.}
\usage{
get_TPMlike_counts(raw_counts)
}
\arguments{
\item{raw_counts}{: raw counts matrix (\emph{peaks} x \emph{samples}).}
}
\value{
a normalized matrix (\emph{peaks} x \emph{samples}).
}
\description{
Normalizing raw ATAC-Seq counts using a TPM-like apporach.
}
\keyword{internal}
