% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_fun.R
\name{match_peaks}
\alias{match_peaks}
\title{Matching the peaks coordinates of the input matrix with the peaks of the ATAC-Seq reference profiles.}
\usage{
match_peaks(bulk_matrix, profile_features, distance = 0)
}
\arguments{
\item{bulk_matrix}{: matrix corresponding to the ATAC-Seq bulk data to deconvolve.}

\item{distance}{(optional, default is 0): minimum distance to consider two regions as overlapping.}
}
\value{
The matrix with lifted coordinates as rownames.
}
\description{
Matching the peaks coordinates of the input matrix with the peaks of the ATAC-Seq reference profiles.
}
\keyword{internal}
