% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_fun.R
\name{run_liftOver}
\alias{run_liftOver}
\title{LiftOver of the peaks coordinates of the input matrix.}
\usage{
run_liftOver(bulk_matrix, from = "hg19")
}
\arguments{
\item{bulk_matrix}{: matrix corresponding to the ATAC-Seq bulk data to deconvolve.}

\item{from}{(optional, default is hg19): the genome version associated to the regions/rownames of bulk_matrix}
}
\value{
The matrix with lifted coordinates as rownames.
}
\description{
If the peaks coordinates of the input matrix are in hg19, they will be
lifted over hg38, which correspond to the genome used for the generation of
ATAC-Seq reference profiles.
}
\keyword{internal}
