% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_descr.R
\docType{data}
\name{TRef}
\alias{TRef}
\title{Reference profiles obtained from single cell data of tumor infiltrating
cells.}
\format{
A list of 3 elements: \describe{ \item{$refProfiles,
  $refProfiles.var}{Matrices (nGenes x nRefCells) of the gene expression (in
  TPM counts) from the reference cells and the variability of this gene
  expression for each gene and each cell type} \item{$sigGenes}{A list of
  signature genes used to deconvolve the cell proportions} }
}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE72056}
}
\usage{
data(TRef)
}
\description{
A dataset containing the reference profiles obtained from various
tumor infiltrating non-malignant cell types: \emph{B cells},
\emph{cancer-associated fibroblasts}, \emph{CD4 T cells}, \emph{CD8 T cells},
\emph{endothelial cells}, \emph{Macrophages} and \emph{NK cells}.
}
\details{
These were obtained from single-cell RNA-seq data from 9 donors from
the publication of \href{http://science.sciencemag.org/content/352/6282/189.full}{
\cite{Tirosh et al., 2016, Science}}. The samples
come from melanoma tumors (extracted from primary tumors and non-lymphoid
tissue metastases). The classification for each sample with
respect to each cell type is the one given by Tirosh et al., except for
the CD4 T cells and CD8 T cells, that were identified from the T cells based
on the expression of CD4, CD8A and CD8B as described in
\href{https://elifesciences.org/articles/26476}{
 \cite{Racle et al., 2017, eLife}}.
}
\keyword{datasets}
