% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_descr.R
\docType{data}
\name{atacRef_PBMC}
\alias{atacRef_PBMC}
\title{ATAC-Seq reference profiles containing cell-types found in peripheral blood mononuclear cells (PBMCs).}
\format{
A list of 3 elements: \describe{ \item{$refProfiles,
  $refProfiles.var}{Matrices (nFeatures x nRefCells) of the chromatin accessibility (in
  TPM-like counts) from the reference cells and the variability of this chromatin
  accessibility for each peak and each cell type} \item{$sigPeaks}{A list of
  marker peaks used to deconvolve the cell proportions} }
}
\usage{
atacRef_PBMC
}
\description{
A dataset containing the reference profiles obtained from immune cell
samples of \emph{B cells}, \emph{CD4 T cells}, \emph{CD8 T cells},
\emph{Monocytes}, \emph{NK cells}, \emph{Dendritic cells} and \emph{Neutrophils}, purified from
PBMC or whole blood.
}
\details{
The original samples were obtained from healthy donors and donors ......
}
\keyword{datasets}
