% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_descr.R
\docType{data}
\name{atacRef_TME}
\alias{atacRef_TME}
\title{ATAC-Seq reference profiles containing cancer-relevant cell-types.}
\format{
A list of 3 elements: \describe{ \item{$refProfiles,
  $refProfiles.var}{Matrices (nGenes x nRefCells) of the chromatin accessibility (in
  TPM-like counts) from the reference cells and the variability of this chromatin
  accessibility for each peak and each cell type} \item{$sigGenes}{A list of
  marker peaks used to deconvolve the cell proportions} }
}
\usage{
atacRef_TME
}
\description{
A dataset containing the reference profiles obtained from immune cell
samples of \emph{B cells}, \emph{CD4 T cells}, \emph{CD8 T cells},
\emph{Macrophages}, \emph{NK cells}, \emph{Dendritic cells}, \emph{Neutrophils}, \emph{fibroblasts} and \emph{endothelial cells}
purified from PBMC or whole blood or .......
}
\details{
The collected cells from healthy or diseased samples and stimulated cells.
}
\keyword{datasets}
