% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_descr.R
\docType{data}
\name{liftover_chains}
\alias{liftover_chains}
\title{Chains to lift over hg19 or hg18 coordinates to hg38 coordinates.}
\format{
Chain object: \describe{ \item{liftover_chains}{Chain objects to lift over hg19 or hg18 coordinates to hg38} }
}
\source{
\enumerate{
 \item \url{https://hgdownload.cse.ucsc.edu/goldenpath/hg19/liftOver/hg19ToHg38.over.chain.gz},
 \item \url{https://hgdownload.cse.ucsc.edu/goldenpath/hg18/liftOver/hg18ToHg38.over.chain.gz}
 }
}
\usage{
liftover_chains
}
\description{
Chains object to perform liftover
}
\keyword{datasets}
